+++ WalletWasabi.Fluent/ViewModels/Login/FullScreenLoginViewModel.cs
@@ -1,15 +1,19 @@
-﻿using WalletWasabi.Fluent.ViewModels.Navigation;
+﻿using System.Collections.ObjectModel;
+using System.Linq;
+using WalletWasabi.Fluent.ViewModels.Wallets;
 using WalletWasabi.Wallets;
 
 namespace WalletWasabi.Fluent.ViewModels.Login
 {
-	public partial class FullScreenLoginViewModel : RoutableViewModel
+	public class FullScreenLoginViewModel : LoginViewModelBase
 	{
-		public FullScreenLoginViewModel(WalletManager walletManager)
+		public FullScreenLoginViewModel(WalletManager walletManager, ObservableCollection<WalletViewModelBase> wallets)
+			: base(walletManager)
 		{
-			// Login = new LoginViewModel(wallet, walletManager);
+			Wallets = wallets;
+			SelectedWallet = wallets.First();
 		}
 
-		public LoginViewModel Login { get; }
+		public ObservableCollection<WalletViewModelBase> Wallets { get; }
 	}
 }
