+++ WalletWasabi.Fluent/Views/Login/LoginView.axaml
@@ -2,36 +2,17 @@
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             xmlns:c="clr-namespace:WalletWasabi.Fluent.Controls"
              xmlns:login="clr-namespace:WalletWasabi.Fluent.ViewModels.Login"
              mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
              x:DataType="login:LoginViewModel"
              x:CompileBindings="True"
              x:Class="WalletWasabi.Fluent.Views.Login.LoginView">
-  <StackPanel HorizontalAlignment="Stretch"
-              VerticalAlignment="Center">
-
-    <Viewbox MaxHeight="200">
-      <Image Source="/Assets/HardwareIcons/ledger.png" Margin="10"/>
-    </Viewbox>
-
-    <Panel Height="150" MaxWidth="350">
-
-      <!-- Enter password -->
-      <StackPanel IsVisible="{Binding !IsPasswordIncorrect}" VerticalAlignment="Bottom">
-        <Label Content="Password:"/>
-        <TextBox Text="{Binding Password}" Watermark="type here..." PasswordChar="•" Classes="revealPasswordButton"/>
-        <Panel Margin="0 -20 0 0 ">
-          <Button Classes="activeHyperLink skip" Content="Forgot password?" HorizontalAlignment="Left"/>
-          <Button Content="Login" IsDefault="True" Classes="action" HorizontalAlignment="Right" Command="{Binding NextCommand}"/>
-        </Panel>
-      </StackPanel>
-
-      <!-- Incorrect password -->
-      <StackPanel Spacing="20" IsVisible="{Binding IsPasswordIncorrect}" VerticalAlignment="Bottom">
-        <TextBlock Text="The password is incorrect! Try Again." HorizontalAlignment="Center"/>
-        <Button Content="OK" Classes="action" HorizontalAlignment="Center" Command="{Binding OkCommand}"/>
-      </StackPanel>
-
-    </Panel>
-  </StackPanel>
+  <c:ContentArea Title="Login"
+                 Caption="Enter your password to access your wallet"
+                 HorizontalAlignment="Stretch"
+                 ScrollViewer.VerticalScrollBarVisibility="Disabled"
+                 IsBusy="{Binding IsBusy}">
+      <c:LoginControl/>
+  </c:ContentArea>
 </UserControl>
