+++ WalletWasabi.Fluent/Views/Login/LoginView.axaml.cs
@@ -1,5 +1,4 @@
 using Avalonia.Controls;
-using Avalonia.Interactivity;
 using Avalonia.Markup.Xaml;
 
 namespace WalletWasabi.Fluent.Views.Login
