--- WalletWasabi.Fluent/Views/Login/SimpleLoginView.axaml.cs
@@ -1,18 +1,0 @@
-using Avalonia.Controls;
-using Avalonia.Markup.Xaml;
-
-namespace WalletWasabi.Fluent.Views.Login
-{
-	public class SimpleLoginView : UserControl
-	{
-		public SimpleLoginView()
-		{
-			InitializeComponent();
-		}
-
-		private void InitializeComponent()
-		{
-			AvaloniaXamlLoader.Load(this);
-		}
-	}
-}
